# website/public/openapi.yaml
openapi: 3.1.0
info:
  title: SageReasoning API
  version: 3.0.0
  description: |
    Structured reasoning evaluation API based on Stoic philosophy.

    At its core is **sage-reason** — a universal reasoning layer that applies
    the Stoic core triad (Control Filter, Passion Diagnosis, Oikeiosis) to any
    decision input. All skills and wrappers call sage-reason internally.

    All responses use a standard envelope: `{ result, meta }`.

    Ancient reasoning, modern application. Does not consider legal, medical,
    financial, or personal obligations.
  contact:
    name: SageReasoning
    url: https://sagereasoning.com
  license:
    name: SageReasoning Proprietary Licence
    url: https://github.com/quartermileclint/stoic-brain/blob/main/LICENSE
servers:
  - url: https://www.sagereasoning.com
    description: Production
...
paths:
...
  /api/evaluate:
    post:
      tags: [demo]
      summary: Instant demo — no auth required
      description: |
        **Outcome:** Try sage-reason instantly — no signup, no API key.
        **Cost + Speed:** Free, ~2s. Quick depth only.
        **Chains To:** /api/reason (with auth for deeper analysis).

        500 character input limit. 5 requests/minute rate limit.
      operationId: evaluate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input]
              properties:
                input:
                  type: string
                  maxLength: 500
                  description: The decision to evaluate.
      responses:
        '200':
          description: Quick evaluation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeResponse'
...
  /api/score-iterate:
    post:
      tags: [deliberation]
      summary: Start or continue a deliberation chain
      description: |
        **Outcome:** Iterative decision refinement — submit, get feedback, revise, track improvement.
        **Cost + Speed:** ~$0.033, ~2s per iteration.
        **Chains To:** sage-score.
      operationId: scoreIterate
      security:
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                context:
                  type: string
                agent_id:
                  type: string
                chain_id:
                  type: string
                revised_action:
                  type: string
                revision_rationale:
                  type: string
      responses:
        '200':
          description: Chain step result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeResponse'
...
  /api/reflect:
    post:
      tags: [evaluation]
      summary: End-of-day decision review
      description: |
        **Outcome:** Identifies patterns in daily decisions + structured Stoic examination.
        **Cost + Speed:** ~$0.033, ~2s.
        **Chains To:** —
      operationId: reflect
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actions]
              properties:
                actions:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Reflection result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Supabase JWT or API key (sr_live_...)
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key (sr_live_...)
  schemas:
...
    ResponseMeta:
      type: object
      properties:
        endpoint:
          type: string
        ai_model:
          type: string
        ai_generated:
          type: boolean
        latency_ms:
          type: integer
        cost_usd:
          type: number
          nullable: true
        is_deterministic:
          type: boolean
        evaluated_at:
          type: string
          format: date-time
        composability:
          type: object
          properties:
            next_steps:
              type: array
              items:
                type: string
            recommended_action:
              type: string
        usage:
          type: object
          properties:
            monthly_calls_used:
              type: integer
            monthly_limit:
              type: integer
            monthly_remaining:
              type: integer
            usage_trigger:
              type: string
              enum: [pre_limit, at_limit]
              nullable: true
            value_summary:
              type: string
...
    EnvelopeResponse:
      type: object
      description: Standard response envelope used by all endpoints.
      properties:
        result:
          type: object
          description: Evaluation-specific result data.
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...